requirements.txt
pandas
numpy
pyyaml
kiteconnect
talib-binary
openpyxl
kiteconnect


